Added the ability to store install location and locale in a .txt file.
If the program is compiled in a .jar, it should create a properties.txt file in the same directory as the .jar.

If the program is not in a .jar, it will try to make a the properties.txt file in the same directory as the compiled class files.

So, it should work for both debugging purposes and distribution purposes as well, hopefully, at least on linux.